user_manager/requirements.txt
flask
mysql-connector-python
grpcio
grpcio-tools
flask[async]==2.2.5